Created show product view.
Still missing comment generation. Structure and relationships for comment entity done. now a user has several comments made to him each of which has an author mapped to the user object of the author.

The rest is pretty much done, its still a bit aesthetic but so its the rest of the web xD. Medals are put instead of products bought, logic works perfect. Thre rest of info fields are correct, just missing  buy form. Buy form should have a quantity and a buy button.